==> Scripts/1 -  Instacart Data import and descriptive analysis.ipynb <==
# Data Import and Descriptive Analysis

# Contents

## 1- Importing Libraries
## 2- Importing Data
## 3- Finding Maixmum Value for aisle_id Column
## 4- Finding the Data Type for the department_id column

## 1) Importing Libraries

import os
import pandas as pd
import numpy as np

## 2) Importing Data

path = r'/Users/calebhill/09:01:2023 - Instacart Basket Analysis'

path 

vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

vars_list

df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), usecols = vars_list)

df.head()

df_prod = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

print(df_prod.head(20))
print(df_prod.tail(35))

print(df_prod.columns)

print(df_prod.shape)

df_prod.describe()

## 3) Finding Maixmum Value for aisle_id Column


Maximum Value: 134.0

df_prod.dtypes

## 4) Finding the Data Type for the department_id column

Datatype: int64
