==> Scripts/2 - Data Wrangling and Subsetting .ipynb <==
# Data Wrangling and Subsetting

## Contents

1- Wrangling Procedures

2- Changing Numerical Data Types

3- Changing an Unintuitive Column Name

4- Finding the Busiest Hours for Placing Orders

5- Determing the Value of '4' in the department_id Column

6- Information about Breakfast Item Sales

7- Finding Details Related to Dinner Parties

8- Finding Counts for the New Dinner Party DataFrame

9- All Information about user_id: '1'

10- Export orders_wrangled.csv DataFrame

11- Export the df_dep_t_new dataframe as “departments_wrangled.csv” 

import pandas as pd
import os
import numpy as np

path = r'/Users/calebhill/09:01:2023 - Instacart Basket Analysis'

df_prod = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'), index_col = False)

## 1) Wrangling Procedures

#removes column
df_ords = df_ords.drop(columns = ['eval_set'])

#provides the count of each variable in days_since_prior_order column
df_ords['days_since_prior_order'].value_counts(dropna = False)

#renames column
df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

df_ords.head()

#Changes Datatype for order_id
df_ords['order_id'] = df_ords['order_id'].astype('str')

#confirms that datatype has been changed
print(df_ords['order_id'].dtype)

#Transposes df_dep
df_dep_t = df_dep.T

df_dep_t

#adds an index
df_dep_t.reset_index()

#takes the first row or dp_dep_t as the header
new_header = df_dep_t.iloc[0]

new_header

#creates a copy of all the rows after one of the data frame
df_dep_t_new = df_dep_t[1:]

df_dep_t_new

#adds the header to the columns in the dataframe
df_dep_t_new.columns = new_header

df_dep_t_new

df_prod.head()

#turns df_dep_t_new into a data dictionary
data_dict = df_dep_t_new.to_dict('index')

data_dict

print(data_dict.get('19'))

df_ords.head()

## 2) Changing Numerical Data Types

#changes the column: order_number to a string
df_ords['user_id'] = df_ords['user_id'].astype('str')

#confirms that the datatype is object
print(df_ords['user_id'].dtype)

## 3) Changing an Unintuitive Column Name

df_ords.rename(columns = {'days_since_prior_order' : 'days_since_last_order'}, inplace = True)

## 4) Finding the Busiest Hours for Placing Orders

#finds the hour where most orders are placed
df_ords['order_hour_of_day'].value_counts(dropna = False).head(1)

#the busiest hour for placing orders is 10

## 5) Determing the Value of "4" in the department_id Column

#prints the value of 4 in the data dictionary
print(data_dict.get('4')['department'])

## 6) Information about Breakfast Item Sales.

df_breakfast_prods = df_prod.loc[df_prod['department_id'].isin([14])]

df_breakfast_prods.head()

## 7) Finding Details Related to Dinner parties

#produces a subset with only alcohol, deli, beverages and meat/seafood products; sorted by department_id
df_dinner_party_prods = df_prod.loc[df_prod['department_id'].isin([5,7,12,20])].sort_values('department_id')

df_dinner_party_prods.head()

## 8) Finding Counts for the New Dinner Party DataFrame

#provides the row count of df_dinner_party_prods
print(f'{df_dinner_party_prods.shape[0]} rows in the dinner party dataframe')

## 9) All information about user_id: "1"

#pulls the statistacl information on user_id: 1
df_ords.loc[df_ords['user_id']=='1'].describe()

## 10) Export orders_wrangled.csv data frame.

df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_wrangled.csv'))

## 11) Export the df_dep_t_new dataframe as “departments_wrangled.csv” 

df_dep_t_new.to_csv(os.path.join(path, '02 Data','Prepared Data', 'departments_wrangled.csv'))
