==> Scripts/4A - Combining and Exporting Data.ipynb <==
# Combining and Exporting Data: A

## Contents

1- Merging Prepared Orders Data with the New orders_product_prior DataFrame¶

2- Exporting the Merged File in Pickle Format as “orders_products_combined.pkl”

import pandas as pd
import os
import numpy as np

path = r'/Users/calebhill/09:01:2023 - Instacart Basket Analysis'

df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders_products_prior.csv'))
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_checked.csv'), index_col = 0)
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_checked.csv'), index_col = 0)

## 1) Merging Prepared Orders Data with the New orders_product_prior DataFrame

df_ords_prior.value_counts(dropna = False)

df_ords_prior.head()

df_ords_prior.shape

df_ords.shape

df_ords.head()

#merges the two dataframes using the order_id column
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

df_merged_large.head()

df_merged_large['_merge'].value_counts()

df_merged_large.shape

## 2) Exporting the Merged File in Pickle Format as “orders_products_combined.pkl”

df_merged_large.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_combined.pkl'))
