==> Scripts/4B - Combining and Exporting Data.ipynb <==
# Combining and Exporting Data: B

## Contents

1- Importing and Performing Basic Checks on the Data

2- Merging the df_ords_prods_combined and df_prods DataFrames

3- Exporting the Newly Merged DataFrame

import pandas as pd
import numpy as np
import os

## 1) Importing and Performing Basic Checks on the Data

path = r'/Users/calebhill/09:01:2023 - Instacart Basket Analysis'

df_ords_prods_combined = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_checked.csv'), index_col = 0)

#checks the dimensions of the dataframe
df_ords_prods_combined.shape

df_ords_prods_combined.head()

df_prods.head()

## 2) Merging the df_ords_prods_combined and df_prods DataFrames

#merges the two dataframes
orders_products_merged = df_prods.merge(df_ords_prods_combined, on = 'product_id', indicator = 'exists')

#confirms everything merged well
orders_products_merged.head()

## 3) Exporting the Newly Merged DataFrame

#exports the newly merged dataframe as a pickle file into Prepared data
orders_products_merged.to_pickle(os.path.join(path, '02 Data','Prepared Data', 'orders_products_merged.pkl'))
