==> Scripts/4.5 Exercise.ipynb <==
# 4.5 - Data Consistency Checks

## Contents

1- Performing Consistency Checks of the df_prods DataFrame 

2- Interpreting Statistical Information from the df_ords DataFrame

3- Checking for Mixed-Type Data in the df_ords DataFrame.

4- Running a Check for Missing Values in the df_ords dataframe and Reporting on it

5- Running a Check for Duplicate Values in the df_ords data and Reporting on it

6- Exporting the Cleaned df_prods and df_ords Data as “.csv” files in the “Prepared Data” Folder

import os
import pandas as pd
import numpy as np

path = r'/Users/calebhill/09:01:2023 - Instacart Basket Analysis'

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'), index_col = 0)
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = 0)

## 1) Performing Consistency Checks of the df_prods DataFrame

df_prods.isnull().sum()

df_prods.head()

df_nan = df_prods[df_prods['product_name'].isnull() == True]

df_nan

df_prods.shape

df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

df_prods_clean.shape

df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups

df_prods_clean.shape

df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

df_prods_clean_no_dups.shape

df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data','Prepared Data', 'products_checked.csv'))

## 2) Interpreting Statistical Information from the df_ords DataFrame

df_ords.describe()

## It looks like the order_hour_of_day column has a max value of 2.3 when it should be 24. 

## 3) Checking for Mixed-Type Data in the df_ords DataFrame.

#checks for mixed-type data by looping through each column with the provided parameters to check for inconsitencies
for col in df_ords.columns.tolist():
    weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
    if len (df_ords[weird]) > 0:
        print(f'There are mixed values in column:{col}')
    else:
        print(f'No mixed values in column: {col}')

## 4) Running a Check for Missing Values in the df_ords dataframe and Reporting on it

df_ords.isnull().sum() #checks for missing values in the dataframe

df_ords_nan = df_ords[df_ords['days_since_last_order'].isnull()== True] #puts NaN values and there corresponding columns in a subset

df_ords_nan #prints the subset 

## There are 206,209 missing values in the days_since_last_order column. It seems reasonable to assume that these are first time customers and if you look at the order number for each corresponding order_number column, the value is 1 

## Because this data is consistent with what you would expect for a first time customer I will leave the values as they are.

## 5) Running a Check for Duplicate Values in the df_ords data and Reporting on it


df_ords_dups = df_ords[df_ords.duplicated()] #puts any duplicates into a subset

df_ords_dups #looks at the subset to check if any duplicates were found

## No Duplicates Found

## 6) Exporting the Cleaned df_prods and df_ords Data as “.csv” files in the “Prepared Data” Folder 

#df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data','Prepared Data', 'products_checked.csv'))
#(I already did this during the reading)

df_ords.to_csv(os.path.join(path, '02 Data','Prepared Data', 'orders_checked.csv'))
