==> Scripts/5 - Deriving New Variables.ipynb <==
# Deriving New Variables

## Contents

1- Creating price_label and busiest_day Columns

2- Creating busiest_days Column

3- Creating busiest_period_of_day Column

4- Printing the frequency of busiest_period_of_day Column

5- Exporting the Updated dataframe as a Pickle File and Storing it in the Prepared Data Folder

import numpy as np
import pandas as pd
import os

path = r'/Users/calebhill/09:01:2023 - Instacart Basket Analysis/02 Data/Prepared Data'

filename = 'orders_products_merged.pkl'

filepath = os.path.join(path, filename)

ords_prods_merge = pd.read_pickle(filepath)

#creates a new dataframe with the first million rows
df = ords_prods_merge[:1000000]

## 1) Creating price_label and busiest_day Columns

#A function that categorizes the product prices into High, Mid, and Low ranges
def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

df['price_range'] = df.apply(price_label, axis = 1) #calls the function with the dataframe

df['price_range'].value_counts(dropna = False) #confirms the data was inputted correctly

df['prices'].max() #confirms the max price

I'm not sure why I'm getting a higher number and I'm not sure how to sort the first 1000000 rows to match what the exercise had.

df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'

Do you know why I'm getting this warning?

#The next three cells replicate the the previous user function but by using the loc function
df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df['price_range_loc'].value_counts(dropna = False)

#the next three cells replicate the same functions as the previous three but for the whole dataframe
ords_prods_merge.loc[ords_prods_merge['prices'] >15, 'price_range_loc'] = 'High Range Product'

ords_prods_merge.loc[(ords_prods_merge['prices']<=15) & (ords_prods_merge['prices']>5), 'price_range_loc'] = 'Mid Range Product'

ords_prods_merge.loc[ords_prods_merge['prices'] <=5, 'price_range_loc'] = 'Low Range Product'

ords_prods_merge['price_range_loc'].value_counts(dropna = False) #confirms the column was inputted correctly

ords_prods_merge['orders_day_of_week'].value_counts(dropna = False) #displays the order count for each day of the week

#a for loop that categorizes the days of the week by Busiest, Least Busy and regular and outputs a list
result = [] #creates an empty list for the values

for value in ords_prods_merge["orders_day_of_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

#creates a busiest day column using the list created in the previous cell
ords_prods_merge['busiest_day'] = result

#confirms the column was inputted correctly
ords_prods_merge['busiest_day'].value_counts(dropna = False)

ords_prods_merge.head()

#renamed the column
ords_prods_merge.rename(columns ={'price_range_loc':'price_label'}, inplace = True)

ords_prods_merge.head()

## 2) Creating busiest_days Column

#a for loop that categorizes the days of the week by busiest, second busiest, slowest, second slowest and regular and outputs a list
new_list = [] #creates an empty list for the day values

for day in ords_prods_merge['orders_day_of_week']:
    if day == 0:
        new_list.append('Busiest Day')
    elif day == 1:
        new_list.append('Second Busiest Day')
    elif day == 4:
        new_list.append('Slowest Day')
    elif day == 3:
        new_list.append('Second Slowest Day')
    else:
        new_list.append('Regularly Busy')

#creates a new column with the list outputted by the previous for loop
ords_prods_merge['busiest_days'] = new_list

#confirms the column was inputted correctly
ords_prods_merge['busiest_days'].value_counts(dropna = False)

Everything looks good

## 3) Creating busiest_period_of_day Column

#displays the count of orders for each hour of the day
ords_prods_merge['order_hour_of_day'].value_counts(dropna = False)

#creates an empty list to put the values created by the for loop
hour_list = []
#I divided the hours of the day into the three categories based on the count of their orders
most_orders =    [10,11,14,15,13,12,16,9]
average_orders = [17,8,18,19,20,7,21,22]
least_orders =   [23,6,0,1,5,2,4,3]

for hour in ords_prods_merge['order_hour_of_day']:
    if hour in most_orders:
        hour_list.append('Most orders')
    elif hour in average_orders:
        hour_list.append('Average orders')
    elif hour in least_orders:
        hour_list.append('Least orders')

#puts the values into a column in the df
ords_prods_merge['busiest_period_of_day'] = hour_list

## 4) Printing the frequency of busiest_period_of_day Column

ords_prods_merge['busiest_period_of_day'].value_counts(dropna = False)

#double checking everything looks good
ords_prods_merge.head()

##  5) Exporting the Updated dataframe as a Pickle File and Storing it in the Prepared Data Folder

ords_prods_merge.to_pickle(os.path.join(path, 'orders_products_merged_2.0.pkl'))
