==> Scripts/6 - Grouping Data & Aggregating Variables.ipynb <==
# Grouping Data & Aggregating Variables

## Contents

1- Finding the Mean of each order_number Value Grouped by Department ID

2- Creating a Loyalty Flag for Customers

3- Creating a Loyalty Flag for Customers based on Customer's max_order

4- Creating a Spending Flag based on Customer's average_price

5- Creating Order Frequency Flag based on Customer's median_frequency

6- Exporting the DataFrame

import pandas as pd
import numpy as np
import os

path = r'/Users/calebhill/09:01:2023 - Instacart Basket Analysis/02 Data/Prepared Data'

filename = 'orders_products_merged_2.0.pkl'

filepath = os.path.join(path, filename)

ords_prods_merge = pd.read_pickle(filepath)

ords_prods_merge['max_order'] = ords_prods_merge.groupby(['user_id'])['order_number'].transform(np.max)

## 1) Finding the Mean of each order_number Value Grouped by Department ID

#outputs the average order number for each department
ords_prods_merge.groupby(['department_id']).agg({'order_number':['mean']})

## 2) Creating a Loyalty Flag for Customers based on Customer's max_order
- If max_order is greater than 40, then: Loyal Customer
- If max_order is less than or equal to 40 and greater than 10, then: Regular Customer
- If max_order is less than or equal to 10, then: New Customer

#the next three cells create the flag based on the criteria
ords_prods_merge.loc[ords_prods_merge['max_order'] > 40, 'loyalty_flag'] = 'Loyal customer'

ords_prods_merge.loc[(ords_prods_merge['max_order'] <= 40) & (ords_prods_merge['max_order'] > 10), 'loyalty_flag'] = 'Regular customer'

ords_prods_merge.loc[ords_prods_merge['max_order'] <= 10, 'loyalty_flag'] = 'New customer'

## 3) Getting the Basic Statistics for the New Loyalty Flag

#turns the outliers over 100 into NaN
ords_prods_merge.loc[ords_prods_merge['prices'] >100, 'prices'] = np.nan

#outputs some basic statistic for the new loyalty_flag column
print(ords_prods_merge.groupby(['loyalty_flag']).agg({'prices': ['mean', 'max', 'min','std']}))

## 4) Creating a Spending Flag based on Customer's average_price
- If average_price is less than 10, then: Low Spender
- If average_price is greater than or equal to 10, then: High Spender

#the next three cells create the flag based on the criteria
ords_prods_merge['average_price'] = ords_prods_merge.groupby(['user_id'])['prices'].transform(np.mean)

ords_prods_merge.loc[ords_prods_merge['average_price'] < 10, 'spending_flag'] = 'Low spender'

ords_prods_merge.loc[ords_prods_merge['average_price'] >= 10, 'spending_flag'] = 'High spender'

#confirms the output is as expected
ords_prods_merge[['user_id', 'average_price', 'spending_flag']].head(100)

#checks frequency of the new column
ords_prods_merge['spending_flag'].value_counts(dropna = False)

## 5) Creating Order Frequency Flag based on Customer's median_frequency
- If the median of “days_since_prior_order” is higher than 20, then the customer should be labeled a “Non-frequent customer.”

- If the median is higher than 10 and lower than or equal to 20, then the customer should be labeled a “Regular customer.”

- If the median is lower than or equal to 10, then the customer should be labeled a “Frequent customer.”

#creates a median frequency column to use when creating the flag
ords_prods_merge['median_frequency'] = ords_prods_merge.groupby(['user_id'])['days_since_last_order'].transform(np.median)

#the next 3 cells create the flag based on the given criteria
ords_prods_merge.loc[ords_prods_merge['median_frequency'] > 20, 'order_frequency_flag'] = 'Non-frequent customer'

ords_prods_merge.loc[(ords_prods_merge['median_frequency'] >10) & (ords_prods_merge['median_frequency'] <= 20), 'order_frequency_flag'] = 'Regular Customer'

ords_prods_merge.loc[ords_prods_merge['median_frequency'] <= 10, 'order_frequency_flag'] = 'Frequent customer'

#confirms results are as expected
ords_prods_merge[['user_id', 'median_frequency', 'order_frequency_flag']].head(2000)

ords_prods_merge['order_frequency_flag'].value_counts(dropna = False)

ords_prods_merge[ords_prods_merge['order_frequency_flag'].isnull()][['user_id', 'days_since_last_order', 'median_frequency', 'order_frequency_flag', 'order_number']]

These are all first time orders so the missing values are expected

## 6) Exporting the DataFrame

ords_prods_merge.to_pickle(os.path.join(path, 'orders_products_merged_3.0.pkl'))
