==> Scripts/7A - Data Visualization with Python.ipynb <==
# Data Visualization with Python: A

## Contents

1- Importing Analysis Libraries and the Customer Data Set

2- Wrangling the Customer Data Set

3- Performing Quality and Consistency Checks on the Customer Data Set

4- Combining Customer Data Set with the Prepared Instacart Data

5- Exporting the New DataFrame as a Pickle File

## 1) Importing Analysis Libraries and the Customer Data Set

#imports libraries
import os
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

path = r'/Users/calebhill/09:01:2023 - Instacart Basket Analysis/02 Data/Original Data'

path_2 = r'/Users/calebhill/09:01:2023 - Instacart Basket Analysis/02 Data/Prepared Data'

filename = 'customers.csv'

filename_2 = 'orders_products_merged_3.0.pkl'

filepath = os.path.join(path, filename)

filepath_2 = os.path.join(path_2, filename_2)

df_cust = pd.read_csv(filepath)

ords_prods_merge = pd.read_pickle(filepath_2)

## 2) Wrangling the Customer Data Set

#a quick look at the data columns and values
df_cust.head()

#changes First Name to first_name
df_cust.rename(columns = {'First Name' : 'first_name'}, inplace = True)

#changes Surnam to last_name
df_cust.rename(columns = {'Surnam' : 'last_name'}, inplace = True)

#changes Gender to gender
df_cust.rename(columns = {'Gender' : 'gender'}, inplace = True)

#changes the Surnam column to last_name
df_cust.rename(columns = {'Age' : 'age'}, inplace = True)

#changes the all caps STATE column just to state
df_cust.rename(columns = {'STATE' : 'state'}, inplace = True)

#changes n_dependants to number_of_dependants for clarity
df_cust.rename(columns = {'n_dependants' : 'number_of_dependants'}, inplace = True)

#changes fam_status to marital_status for clarity
df_cust.rename(columns = {'fam_status' : 'marital_status'}, inplace = True)

#confirms changes
df_cust.head()

#checks the data types for each column
df_cust.dtypes

#change user_id to string type
df_cust['user_id'] = df_cust['user_id'].astype('str')

#confirms the type change
df_cust['user_id'].dtype

## No transposing needed for this file and I'm not sure which columns I can drop yet although I'm pretty sure marital_status and number_of_dependents won't be needed

## 3) Performing Quality and Consistency Checks on the Customer Data Set

df_cust.describe()

#checks for null values
df_cust.isna().sum()

#replaces all NaN values with Unknown
df_cust['first_name'].fillna('Unknown', inplace=True)

#checks for mixed data types
for col in df_cust.columns.tolist():
  weird = (df_cust[[col]].applymap(type) != df_cust[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_cust[weird]) > 0:
    print (col)

#checks for duplicates
df_dups = df_cust[df_cust.duplicated()]

df_dups.head()
#no duplicates found

df_cust.shape

df_cust.to_pickle(os.path.join(path_2, 'customers_checked.pkl'))

## 4) Combining Customer Data Set with the Prepared Instacart Data

#changes the datatype of user_id to string so it matches with the user_id datatype I need to merge on
ords_prods_merge['user_id'] = ords_prods_merge['user_id'].astype('str')

ord_prods_all = df_cust.merge(ords_prods_merge, on = 'user_id')

ord_prods_all.head()

## 5) Exporting the New Dataframe as a Pickle File

ord_prods_all.to_pickle(os.path.join(path_2, 'ords_prods_cust_merged.pkl'))
